Final update
Updated version:
- Implemented all steps mentioned from the previous update: 
     - Implementing the three effects: Extra Boost, Slow Down, and Extra Time.
     - Adding Level 2. Now the NPC cars and effects go faster, but 
     - Adding a main menu scene and a Game Over scene.
- Updated the Development Log and the GDD, check the docs folder.
- Changed the collision code - now effects and NPC cars will check whether they've collided with the player or not. Previous version declared Game Over if any object ever collided.

This will be the final update as the submission date is almost up.

There is a bug, though it has a very small chance of appearing in-game. It involves spawning effects or cars, at which they spawn at the top-left corner of the screen.
Also worth mentioning is that the window size was able to be increased in the previous versions, allowing for out of bounds areas to be seen by the player. This was fixed in this update.